requirements.txt
astroid==2.3.3
colorama==0.4.1
decorators==0.1.1
dj-database-url==0.5.0
autopep8==1.4.4
branca==0.3.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-appconf==1.0.3
django-bootstrap-themes==3.3.6
django-decorators==0.1.1
django-heroku==0.3.1
gunicorn==20.0.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.4
pylint==2.4.3
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==4.1.4
wrapt==1.11.2
django-compressor==2.3
folium==0.10.0
idna==2.8
isort==4.3.21
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.4
pycodestyle==2.5.0
pylint-django==2.0.13
pylint-plugin-utils==0.6
pytz==2019.3
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.1.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.7
wrapt==1.11.2
backcall==0.1.0
beautifulsoup4==4.8.1
bootstrap4==0.1.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
django-bootstrap4==1.0.1
django-extensions==2.2.5
django-imagekit==4.0.2
Faker==2.0.3
get==2019.4.13
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
matplotlib==3.1.2
mpld3==0.3
numpy==1.17.4
pandas==0.25.3
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.2.1
post==2019.4.13
prettytable==0.7.2
prompt-toolkit==2.0.10
public==2019.4.13
pyecharts==1.6.0
Pygments==2.4.2
pyparsing==2.4.5
python-dateutil==2.8.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
simplejson==3.17.0
soupsieve==1.9.4
text-unidecode==1.3
traitlets==4.3.3
wcwidth==0.1.7